requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==3.2.16
djangorestframework==3.14.0
idna==3.4
pydantic==1.10.2
pytz==2022.4
requests==2.28.1
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.12
